fixing offres management nearly done

+++ src/main/java/com/tekup/pfaapisb/Services/OffreEmploiService.java
@@ -7,7 +7,9 @@ import com.tekup.pfaapisb.Repositories.OffreEmploiRepository;
 import com.tekup.pfaapisb.Repositories.RecruteurRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -32,14 +34,13 @@ public class OffreEmploiService {
         offreEmploi.setRecruteur(recruteur);
         return offreEmploiRepository.save(offreEmploi);
     }
-    public Optional<OffreEmploi> getOffreById(Long id) {
-        return offreEmploiRepository.findById(id);
-    }
-    public List<OffreEmploi> getAllOffres() {
-        return offreEmploiRepository.findAll();
-    }
-    public OffreEmploi updateOffre(Long id, OffreEmploiIDTO offreEmploiDTO) {
+
+    public OffreEmploi updateOffre(Long id, OffreEmploiIDTO offreEmploiDTO, String email) {
         OffreEmploi offreEmploi = offreEmploiRepository.findById(id).orElseThrow(() -> new RuntimeException("Offre non trouvée"));
+        // Check if the recruiter owns the offer
+        if (!offreEmploi.getRecruteur().getEmail().equals(email)) {
+            throw new RuntimeException("Vous n'êtes pas autorisé à modifier cette offre.");
+        }
         offreEmploi.setTitre(offreEmploiDTO.getTitre());
         offreEmploi.setDescription(offreEmploiDTO.getDescription());
         offreEmploi.setCompetences(offreEmploiDTO.getCompetences());
@@ -48,7 +49,29 @@ public class OffreEmploiService {
         offreEmploi.setSalaire(offreEmploiDTO.getSalaire());
         return offreEmploiRepository.save(offreEmploi);
     }
-    public void deleteOffre(Long id) {
+
+    @Transactional
+    public void deleteOffreById(Long id, String email) {
+        OffreEmploi offreEmploi = offreEmploiRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Offre non trouvée"));
+
+        // Check if the recruiter owns the offer
+        if (!offreEmploi.getRecruteur().getEmail().equals(email)) {
+            throw new RuntimeException("Vous n'êtes pas autorisé à supprimer cette offre.");
+        }
+
         offreEmploiRepository.deleteById(id);
+    }
+
+    public Optional<OffreEmploi> getOffreById(Long id) {
+        return offreEmploiRepository.findById(id);
+    }
+
+    public List<OffreEmploi> getAllOffresByRecruteur(long id) {
+        return offreEmploiRepository.findByRecruteurId(id);
+    }
+
+    public List<OffreEmploi> getAllOffres() {
+        return offreEmploiRepository.findAll();
     }
 }
